+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # fix-plugin2 Changelog
 
 ## [Unreleased]
+- Added language injection for FIX messages embedded in code strings
+- Added tests for language injection
 
 ## [0.0.1]